revspotify/requirements.txt
eyeD3
requests
spotipy
yt_dlp
youtube_search
python-telegram-bot
python-dotenv
validator_collection
py-deezer
pycryptodome==3.12.0
youtube_search
spotipy